app(variables): pin break-all, and share the tooltip block scan - #1195
Two findings from the review pass on the previous commit.

The layout guards pinned the stack and forgot the wrap. Dropping
`break-all` from `TooltipValue` left every test in the diff green, which
makes the mutation check half a check: the stack is what keeps a long
value at full width, `break-all` is what makes it wrap there rather than
push the tooltip past its cap. Both components now read it off the
rendered element, and dropping it reds three cases.

The new scan was a copy of `tooltip-hint-contrast.test.ts`'s walk, down
to the block regex - the shape this repo has a rule about, since
widening that regex has to reach both guards. The walk moves to
`tooltip-blocks.testkit.ts` and both scans import it.

`app/CLAUDE.md` gains the rule beside the tooltip-contrast one it sits
next to; that list is where a UI rule enforced by a test is written down.
